+++ web/packages/new/photos/components/utils/use-snapshot.ts
